+++ src/components/animatable-cube/animatable-cube.tsx
@@ -44,7 +44,7 @@ export class Cube implements IAnimatableComponent {
   /**
    * Animation manager for Animatable
    */
-  private manager: AnimationManager
+  private manager?: AnimationManager = null
 
   @Element() el!: HTMLElement
 
@@ -325,7 +325,9 @@ export class Cube implements IAnimatableComponent {
    */
   @Method()
   async destroy(): Promise<void> {
-    this.manager.destroyAnimation();
+    if (this.manager !== null) {
+      this.manager.destroyAnimation();
+    }
   }
 
   /**